+++ internal/entrypoint/packet_registry.go
@@ -191,10 +191,10 @@ func registerPackets() {
 	packets.RegisterPacketHandler(22101, []packets.PacketHandler{answer.GetShopStreet})
 	packets.RegisterPacketHandler(16001, []packets.PacketHandler{answer.ShoppingCommandAnswer})
 	packets.RegisterPacketHandler(11501, []packets.PacketHandler{answer.ChargeCommandAnswer})
-	packets.RegisterPacketHandler(11504, []packets.PacketHandler{answer.ChargeConfirmAnswer})
+	packets.RegisterPacketHandler(11504, []packets.PacketHandler{answer.ChargeConfirmCommandAnswer})
 	packets.RegisterPacketHandler(11508, []packets.PacketHandler{answer.ExchangeCodeRedeem})
-	packets.RegisterPacketHandler(11510, []packets.PacketHandler{answer.ChargeFailed})
-	packets.RegisterPacketHandler(11513, []packets.PacketHandler{answer.RefundChargeAnswer})
+	packets.RegisterPacketHandler(11510, []packets.PacketHandler{answer.ChargeFailedCommandAnswer})
+	packets.RegisterPacketHandler(11513, []packets.PacketHandler{answer.RefundChargeCommandAnswer})
 	packets.RegisterPacketHandler(12004, []packets.PacketHandler{answer.RetireShip})
 	packets.RegisterPacketHandler(11401, []packets.PacketHandler{answer.ChatRoomChange})
 	packets.RegisterPacketHandler(50102, []packets.PacketHandler{answer.ReceiveChatMessage})
